Converted a lot of floating point calculations to half precision

+++ ShaderLibrary/GlobalIllumination.hlsl
@@ -4,17 +4,17 @@
 #include "Packages/com.unity.render-pipelines.core/ShaderLibrary/ImageBasedLighting.hlsl"
 #include "Environment.hlsl"
 
-float4 _AmbientLightColor;
+half4 _AmbientLightColor;
 
 struct GlobalIllumination
 {
-	float3 specular;
-	float3 ambient;
+	half3 specular;
+    half3 ambient;
 };
 
-float3 SampleEnvironment(Surface surface, BRDF brdf)
+half3 SampleEnvironment(Surface surface, BRDF brdf)
 {
-	float3 dir = reflect(-surface.viewDirection, surface.normal);
+    half3 dir = reflect(-surface.viewDirection, surface.normal);
 	float mip = PerceptualRoughnessToMipmapLevel(brdf.perceptualRoughness, 6);
 	return SampleSkyboxCubemap(dir, mip).rgb;
 }
