+++ ShaderLibrary/LitInput.hlsl
@@ -10,19 +10,19 @@ TEXTURE2D(_SurfaceTex);
 SAMPLER(sampler_MainTex);
 
 UNITY_INSTANCING_BUFFER_START(UnityPerMaterial)
-	UNITY_DEFINE_INSTANCED_PROP(float4, _Color)
+	UNITY_DEFINE_INSTANCED_PROP(half4, _Color)
 	UNITY_DEFINE_INSTANCED_PROP(float4, _MainTex_ST)
 	UNITY_DEFINE_INSTANCED_PROP(float4, _EmissionTex_ST)
 	UNITY_DEFINE_INSTANCED_PROP(float4, _NormalTex_ST)
 	UNITY_DEFINE_INSTANCED_PROP(float4, _SurfaceTex_ST)
-	UNITY_DEFINE_INSTANCED_PROP(float4, _EmissionColor)
-	UNITY_DEFINE_INSTANCED_PROP(float, _NormalScale)
-	UNITY_DEFINE_INSTANCED_PROP(float, _Cutoff)
-	UNITY_DEFINE_INSTANCED_PROP(float, _ZWrite)
-	UNITY_DEFINE_INSTANCED_PROP(float, _Metallic)
-	UNITY_DEFINE_INSTANCED_PROP(float, _Smoothness)
-	UNITY_DEFINE_INSTANCED_PROP(float, _Fresnel)
-	UNITY_DEFINE_INSTANCED_PROP(float, _IndividualShadowStrength)
+	UNITY_DEFINE_INSTANCED_PROP(half4, _EmissionColor)
+	UNITY_DEFINE_INSTANCED_PROP(half, _NormalScale)
+	UNITY_DEFINE_INSTANCED_PROP(half, _Cutoff)
+	UNITY_DEFINE_INSTANCED_PROP(half, _ZWrite)
+	UNITY_DEFINE_INSTANCED_PROP(half, _Metallic)
+	UNITY_DEFINE_INSTANCED_PROP(half, _Smoothness)
+	UNITY_DEFINE_INSTANCED_PROP(half, _Fresnel)
+	UNITY_DEFINE_INSTANCED_PROP(half, _IndividualShadowStrength)
 #if defined(GRP_LIT_ADDITIONAL_PER_MATERIAL_PROPERTIES)
 	GRP_LIT_ADDITIONAL_PER_MATERIAL_PROPERTIES
 #endif
@@ -59,53 +59,53 @@ InputConfig GetInputConfig(float2 texcoord)
 	return config;
 }
 
-float4 GetDiffuse(InputConfig config)
+half4 GetDiffuse(InputConfig config)
 {
-	float4 color = PER_MATERIAL_PROP(_Color);
+    half4 color = PER_MATERIAL_PROP(_Color);
 	color *= SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, config.texcoord);
 	return color;
 }
 
-float3 GetEmission(InputConfig config)
+half3 GetEmission(InputConfig config)
 {
-	float4 color = PER_MATERIAL_PROP(_EmissionColor);
+    half4 color = PER_MATERIAL_PROP(_EmissionColor);
 #if defined(_HAS_EMISSION_TEXTURE)
 	color *= SAMPLE_TEXTURE2D(_EmissionTex, sampler_MainTex, config.emission_Texcoord);
 #endif
 	return color.rgb;
 }
 
-float3 GetNormal(InputConfig config)
+half3 GetNormal(InputConfig config)
 {
 #if defined(_HAS_NORMAL_TEXTURE)
 	float4 map = SAMPLE_TEXTURE2D(_NormalTex, sampler_MainTex, config.normal_Texcoord);
-	float scale = PER_MATERIAL_PROP(_NormalScale);
+	half scale = PER_MATERIAL_PROP(_NormalScale);
 	return DecodeNormal(map, scale);
 #else
-	return float3(0, 1, 0);
+    return half3(0, 1, 0);
 #endif
 }
 
-float GetAlpha(InputConfig config)
+half GetAlpha(InputConfig config)
 {
 	return GetDiffuse(config).a;
 }
 
-float GetFinalAlpha(float alpha)
+half GetFinalAlpha(half alpha)
 {
 	return PER_MATERIAL_PROP(_ZWrite) ? 1.0 : alpha;
 }
 
-float GetCutoff(InputConfig config)
+half GetCutoff(InputConfig config)
 {
 	return PER_MATERIAL_PROP(_Cutoff);
 }
 
 struct SurfaceInfo
 {
-	float metallic;
-	float smoothness;
-	float fresnel;
+	half metallic;
+	half smoothness;
+    half fresnel;
 };
 
 SurfaceInfo GetSurfaceInfo(InputConfig config)
@@ -115,7 +115,7 @@ SurfaceInfo GetSurfaceInfo(InputConfig config)
 	result.smoothness = PER_MATERIAL_PROP(_Smoothness);
 	result.fresnel = PER_MATERIAL_PROP(_Fresnel);
 #if defined(_HAS_SURFACE_TEXTURE)
-	float3 msf = SAMPLE_TEXTURE2D(_SurfaceTex, sampler_MainTex, config.surface_Texcoord).xyz;
+	half3 msf = SAMPLE_TEXTURE2D(_SurfaceTex, sampler_MainTex, config.surface_Texcoord).xyz;
 	result.metallic *= msf.x;
 	result.smoothness *= msf.y;
 	result.fresnel *= msf.z;
@@ -123,7 +123,7 @@ SurfaceInfo GetSurfaceInfo(InputConfig config)
 	return result;
 }
 
-void ClipAlpha(float alpha, InputConfig config)
+void ClipAlpha(half alpha, InputConfig config)
 {
 	AlphaDiscard(alpha, GetCutoff(config));
 }
