+++ ShaderLibrary/TerrainGrassPass.hlsl
@@ -15,7 +15,7 @@ struct Attributes
 	float2 texcoord : TEXCOORD0;
 	float3 normalOS : NORMAL;
 	float4 tangentOS : TANGENT;
-	float4 color : COLOR;
+    half4 color : COLOR;
 	UNITY_VERTEX_INPUT_INSTANCE_ID
 };
 
@@ -25,7 +25,7 @@ struct Varyings
 	float3 positionWS : TEXCOORD0;
 	float3 normalWS : NORMAL;
 	float2 texcoord : TEXCOORD1;
-	float4 color : COLOR;
+    half4 color : COLOR;
 	FOG_ATTRIBUTE(2)
 	UNITY_VERTEX_INPUT_INSTANCE_ID
 };
@@ -60,7 +60,7 @@ Varyings GrassVertex(Attributes input)
 
 	output.color = input.color;
 #if defined(_COLOR_SPACE_LINEAR)
-	output.color = float4(SRGBToLinear(input.color.rgb), input.color.a);
+	output.color = half4(SRGBToLinear(input.color.rgb), input.color.a);
 #else
 	output.color = input.color;
 #endif
@@ -77,15 +77,15 @@ FragmentOutput GrassFragment(Varyings input)
 {
 	UNITY_SETUP_INSTANCE_ID(input);
 
-	float4 diffuse = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, input.texcoord);
+    half4 diffuse = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, input.texcoord);
 
 	diffuse.a *= input.color.a;
 	AlphaDiscard(diffuse.a, _Cutoff);
 	diffuse.rgb *= input.color.rgb;
 
-	float metallic = 0;
-	float smoothness = 0;
-	float fresnel = 0;
+	half metallic = 0;
+	half smoothness = 0;
+    half fresnel = 0;
 	Surface surface = MakeSurface(diffuse, input.positionWS, input.normalWS, metallic, smoothness, fresnel, input.positionCS.xy);
 
 	BRDF brdf = GetBRDF(surface);
@@ -94,7 +94,7 @@ FragmentOutput GrassFragment(Varyings input)
 
 	APPLY_FOG(input, lit);
 
-	return MakeFragmentOutput(float4(lit, surface.alpha), surface.normal);
+    return MakeFragmentOutput(half4(lit, surface.alpha), surface.normal);
 }
 
 #endif
